retriever/requirements.txt
jsonlines
pytorch-ignite
transformers
sentence-transformers
torch_optimizer
pandas